Fixed bug in conn16b_struct connection structure type

+++ src/utilities.cu
@@ -20,12 +20,12 @@
  *
  */
 
-int IntPow(int x, unsigned int p)
+int64_t IntPow(int64_t x, unsigned int p)
 {
   if (p == 0) return 1;
   if (p == 1) return x;
   
-  int tmp = IntPow(x, p/2);
+  int64_t tmp = IntPow(x, p/2);
   if (p%2 == 0) return tmp * tmp;
   else return x * tmp * tmp;
 }
